requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cmake==3.25.2
colorama==0.4.6
cryptography==37.0.4
decouple==0.0.7
defusedxml==0.7.1
Django==4.1
django-allauth==0.51.0
django-cors-headers==3.13.0
django-dotenv==1.4.2
djangorestframework==3.13.1
idna==3.3
msgpack==1.0.5
mysqlclient==2.1.1
numpy==1.24.2
oauthlib==3.2.0
opencv-python==4.7.0.72
Pillow==9.4.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.11